% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_py_install.R
\name{conda_py_install}
\alias{conda_py_install}
\title{conda_py_install}
\usage{
conda_py_install(env_name, pkgs = list(conda = "python", py = NULL), ...)
}
\arguments{
\item{env_name}{name of conda environment}

\item{pkgs}{named list of conda/py packages to install (default:
list(conda = "python", py = NULL))}

\item{...}{additional arguments passed to \code{\link{install_miniconda}}}
}
\value{
creates conda environment and returns path
}
\description{
combined installation of conda/py packages in one call
}
