% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_miniconda.R
\name{install_miniconda}
\alias{install_miniconda}
\title{install_miniconda}
\usage{
install_miniconda(
  path = reticulate::miniconda_path(),
  update = TRUE,
  force = FALSE
)
}
\arguments{
\item{path}{The path in which Miniconda will be installed. Note that the
installer does not support paths containing spaces. See \link[reticulate]{miniconda_path} for
more details on the default path used by reticulate.}

\item{update}{Boolean; update to the latest version of Miniconda after install?
(default: TRUE)}

\item{force}{Boolean; force re-installation if Miniconda is already installed
at the requested path? (default: FALSE)}
}
\value{
installs Miniconda at requested path (or skips in case already
installed, where reticulate would otherwise throw an error!)
}
\description{
install_miniconda
}
