% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_miniconda.R
\name{miniconda_conda}
\alias{miniconda_conda}
\title{miniconda_conda}
\usage{
miniconda_conda(path)
}
\arguments{
\item{path}{path}
}
\value{
path to Miniconda executable
}
\description{
miniconda_conda
}
\keyword{internal}
