% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_miniconda.R
\name{miniconda_exists}
\alias{miniconda_exists}
\title{miniconda_exists}
\usage{
miniconda_exists(path)
}
\arguments{
\item{path}{path}
}
\value{
TRUE or FALSE
}
\description{
miniconda_exists
}
\keyword{internal}
