% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_django.R
\name{run_django}
\alias{run_django}
\title{run_django}
\usage{
run_django(condaenv, cmd = "runserver", path_manage.py = "tools/manage.py")
}
\arguments{
\item{condaenv}{name/path of conda environment with Django app}

\item{cmd}{command (default: "runserver")}

\item{path_manage.py}{relative/absolute path to "manage.py" (default:
tools/manage.py")}
}
\value{
runs Django command and opens webbrowser (if cmd == "runserver")
}
\description{
run_django
}
